Add files via upload
updated for PM w/Alabama and Texas updates.
Something funny with Texas data - they seemed to have addressed some issues with enrollment of in-person learning reporting, but a lot of districts previously reporting were just dropped from the report or missing values. I inquired with the team that manages the data for texas and will correct if needed.